requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.11.0
boto3==1.40.11
botocore==1.40.11
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.10.2
cron-descriptor==1.4.5
cryptography==45.0.6
daphne==4.2.1
distlib==0.4.0
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.20.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
hyperlink==21.0.0
identify==2.6.15
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==7.0.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pytest==8.4.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytokens==0.3.0
PyYAML==6.0.3
redis==4.6.0
referencing==0.37.0
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.7
s3transfer==0.13.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
tenacity==9.1.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.13
zope.interface==7.2

# Phase 3: Observability Stack
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0
